fixed prototype fml

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: greed <marvin@codam.nl>                      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/28 12:43:56 by greed          #+#    #+#                */
-/*   Updated: 2019/11/06 10:32:50 by greed         ########   odam.nl         */
+/*   Updated: 2019/11/07 12:48:31 by greed         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t		ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i])
